+++ backend/src/main/java/com/ddbb/dingdong/infrastructure/bus/simulator/RouteSegment.java
@@ -1,4 +1,4 @@
-package com.ddbb.dingdong.simulator;
+package com.ddbb.dingdong.infrastructure.bus.simulator;
 
 import org.springframework.data.geo.Point;
