+++ backend/src/main/java/com/ddbb/dingdong/infrastructure/bus/simulator/subscription/subscriber/CancelableSubscriber.java
@@ -1,4 +1,4 @@
-package com.ddbb.dingdong.simulator.subscription.subscriber;
+package com.ddbb.dingdong.infrastructure.bus.simulator.subscription.subscriber;
 
 import java.util.concurrent.Flow;
