+++ backend/src/main/java/com/ddbb/dingdong/infrastructure/bus/simulator/subscription/subscriber/StubConsoleSubscriber.java
@@ -1,4 +1,4 @@
-package com.ddbb.dingdong.simulator.subscription.subscriber;
+package com.ddbb.dingdong.infrastructure.bus.simulator.subscription.subscriber;
 
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.data.geo.Point;
